reg test for scipy layout opt with value
